moved api out of pages
